proton-version-check/requirements.txt
requests>=2.32.0
PyYAML>=6.0.2

